Rename Common to Utility
Fixes MSVC build.
