+++ src/CMakeLists.txt
@@ -1,6 +1,6 @@
 include_directories(.)
 
-add_subdirectory(Common)
+add_subdirectory(Utility)
 add_subdirectory(Video)
 add_subdirectory(Engine)
 add_subdirectory(Editor)
