+++ src/Editor/GUI/Editors/SceneEditor.cpp
@@ -2,7 +2,7 @@
 
 #include <Engine/Hymn.hpp>
 #include <Engine/Util/FileSystem.hpp>
-#include <Common/Log.hpp>
+#include <Utility/Log.hpp>
 #include <imgui.h>
 
 using namespace GUI;
