+++ src/Engine/MainWindow.cpp
@@ -5,7 +5,7 @@
 #include "Manager/Managers.hpp"
 #include "Manager/RenderManager.hpp"
 #include "Util/Input.hpp"
-#include <Common/Log.hpp>
+#include <Utility/Log.hpp>
 
 MainWindow* MainWindow::instance = nullptr;
 void WindowSizeCallback(GLFWwindow* window, int width, int height);
