+++ src/Engine/Util/FileSystem.cpp
@@ -6,7 +6,7 @@
 #include <cctype>
 #include <cstring>
 #include <fstream>
-#include <Common/Log.hpp>
+#include <Utility/Log.hpp>
 
 // Platform-dependent includes.
 #if defined(_WIN32) || defined(WIN32)
